acos-config Code Refactoring with changing 'acos' by 'a10'
